requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-import-export==4.4.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==25.1.0
openpyxl==3.1.5
packaging==26.0
psycopg==3.3.3
python-dotenv==1.2.2
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
whitenoise==6.12.0
